+++ service/qq/client.go
@@ -1,0 +1,35 @@
+package qq
+
+import (
+	"github.com/dtapps/go-library/utils/golog"
+	"github.com/dtapps/go-library/utils/gorequest"
+)
+
+// ClientConfig 实例配置
+type ClientConfig struct {
+	Key string
+}
+
+// Client 实例
+type Client struct {
+	requestClient *gorequest.App // 请求服务
+	config        struct {
+		key string
+	}
+	log struct {
+		status bool             // 状态
+		client *golog.ApiClient // 日志服务
+	}
+}
+
+// NewClient 创建实例化
+func NewClient(config *ClientConfig) (*Client, error) {
+
+	c := &Client{}
+
+	c.config.key = config.Key
+
+	c.requestClient = gorequest.NewHttp()
+
+	return c, nil
+}