+++ service/qq/const.go
@@ -1,0 +1,9 @@
+package qq
+
+const (
+	apiUrl = "https://apis.map.qq.com"
+)
+
+const (
+	LogTable = "qq"
+)