+++ service/wechatopen/wxa.media_check_async.go
@@ -1,4 +1,4 @@
-package wechatminiprogram
+package wechatopen
 
 import (
 	"context"
@@ -49,7 +49,7 @@ func (c *Client) WxaMediaCheckAsync(ctx context.Context, notMustParams ...gorequ
 	// 参数
 	params := gorequest.NewParamsWith(notMustParams...)
 	// 请求
-	request, err := c.request(ctx, fmt.Sprintf(apiUrl+"/wxa/media_check_async?access_token=%s", c.getAccessToken(ctx)), params, http.MethodPost)
+	request, err := c.request(ctx, fmt.Sprintf(apiUrl+"/wxa/media_check_async?access_token=%s", c.GetAuthorizerAccessToken(ctx)), params, http.MethodPost)
 	// 定义
 	var response WxaMediaCheckAsyncResponse
 	err = gojson.Unmarshal(request.ResponseBody, &response)